Implemented the weapon reloading mechanic
Testing using a UI button that "shoots" the weapon and, when the amount of available bullets reaches 0, runs a coroutine that simulates a gun reloading delay. UI text is also updates and, while the gun is reloading, a text saying "RELOAD" is shown in the weapon display in the bottom-right corner.
Code still needs to be commented.