Minimap on level 1 + Incomplete work on pause menu and HUD code fixing
Can't apply this work on main branch yet. Moved the UI scripts from Steven Up prefab to the HUD canvas prefab and I had to rewrite one line of code in other scripts so that I could manually point to the new place of the UI scripts. Because of that, I have to manually adjust some stuff in the Inspector which got lost in the process and I need to redownload the main branch code to see what it is.